[Refactor] 검색창 컴포넌트 분리 #1134
